фикс удаления аватара пользователя

+++ app/Services/UserService.php
@@ -61,7 +61,7 @@ class UserService
         if ($isAvatarChanged) {
             $this->tempFileService->moveFromTmpFolder(TempFile::FOLDER_USER_AVATAR, $user->avatar);
 
-            $this->tempFileService->removeFileFromFolder(TempFile::FOLDER_USER_AVATAR, $oldAvatar);
+            if($oldAvatar) $this->tempFileService->removeFileFromFolder(TempFile::FOLDER_USER_AVATAR, $oldAvatar);
         }
 
         return $user->refresh();
